Move 'create', 'collection' and helpers to private folder
